requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.2
backports.csv==1.0.7
beautifulsoup4==4.8.1
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cheroot==8.2.1
CherryPy==18.4.0
Click==7.0
colorama==0.4.1
feedparser==5.2.1
firebase-admin==3.1.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Session==0.3.1
future==0.18.2
gast==0.3.2
gcloud==0.18.3
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.20.0
google-pasta==0.1.7
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.3
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.14.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
jaraco.functools==2.0
Jinja2==2.10.3
jws==0.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.2
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
mysqlclient==1.4.4
nltk==3.4.5
numpy==1.17.3
oauth2client==4.1.2
Pattern==3.6
pdfminer.six==20191020
portend==2.6
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycryptodome==3.9.0
pylint==2.4.3
Pyrebase4==4.3.0
pyspellchecker==0.5.2
python-docx==0.8.10
python-jwt==2.0.1
pytz==2019.3
pywin32==225
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
scipy==1.3.1
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
tempora==1.14.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
zc.lockfile==2.0
